# apps/api/generated/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qafiyah API",
...
    "contact": {
      "name": "Qafiyah",
      "url": "https://qafiyah.com"
    },
    "license": {
      "name": "CC0 1.0 (data)",
      "identifier": "CC0-1.0"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
    "/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "description": "Every curated collection (al-dawawin) with its poem count. Counts include primary readings only.",
        "operationId": "collections.list",
        "responses": {
          "200": {
            "description": "All collections.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Send back in If-None-Match to get 304 while the data is unchanged"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_PoemCountStats"
                }
              }
            }
          },
          "304": {
            "description": "Not modified: the If-None-Match ETag still matches, so the body is empty.",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable: the database or search index did not answer in time",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds to wait before retrying"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{slug}": {
      "get": {
        "tags": [
          "collections"
        ],
        "description": "A collection by slug, with its poem count.",
        "operationId": "collections.get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "The term's `slug`, from its list.",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z-]*$"
            },
            "example": "almuallaqat"
          }
        ],
        "responses": {
          "200": {
            "description": "The requested collection.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Send back in If-None-Match to get 304 while the data is unchanged"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemEnvelope_PoemCountStats"
                }
              }
            }
          },
          "304": {
            "description": "Not modified: the If-None-Match ETag still matches, so the body is empty.",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            }
          },
          "400": {
            "description": "Input validation failed",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable: the database or search index did not answer in time",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds to wait before retrying"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/eras": {
      "get": {
        "tags": [
          "eras"
        ],
        "description": "Every literary era (al-usur al-adabiyya) with its poem and poet counts, oldest first, with the unknown era last. Counts include primary readings only.",
        "operationId": "eras.list",
        "responses": {
          "200": {
            "description": "All eras.",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "Send back in If-None-Match to get 304 while the data is unchanged"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEnvelope_CountedStats"
                }
              }
            }
          },
          "304": {
            "description": "Not modified: the If-None-Match ETag still matches, so the body is empty.",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds until the window resets"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "headers": {
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "503": {
            "description": "Temporarily unavailable: the database or search index did not answer in time",
            "headers": {
              "retry-after": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                },
                "description": "Seconds to wait before retrying"
              },
              "x-ratelimit-limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests allowed per hour"
              },
              "x-ratelimit-remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests left in the current hour"
              },
              "x-ratelimit-reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix seconds at which the window resets"
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ItemEnvelope_PoemCountStats": {
        "type": "object",
        "description": "The requested item under `data`.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "name",
              "slug",
              "poemsCount"
            ],
            "properties": {
              "name": {
                "type": "string",
                "description": "The term's name in Arabic.",
                "example": "الحكمة"
              },
              "poemsCount": {
                "type": "integer",
                "format": "int32",
                "description": "Number of poems with this term, primary readings only.",
                "example": 22521
              },
              "slug": {
                "type": "string",
                "description": "The term's slug.",
                "example": "alhikma",
                "pattern": "^[a-z][a-z-]*$"
              }
            }
          }
        }
      },
...
      "ListEnvelope_PoemCountStats": {
        "type": "object",
        "description": "A page of items under `data`, with its place in the whole list under `pagination`.",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "name",
                "slug",
                "poemsCount"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "description": "The term's name in Arabic.",
                  "example": "الحكمة"
                },
                "poemsCount": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Number of poems with this term, primary readings only.",
                  "example": 22521
                },
                "slug": {
                  "type": "string",
                  "description": "The term's slug.",
                  "example": "alhikma",
                  "pattern": "^[a-z][a-z-]*$"
                }
              }
            },
            "description": "The items of this page."
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination",
            "description": "Where this page sits in the list."
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "instance",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "A stable machine-readable code, such as `NOT_FOUND`, `BAD_REQUEST`, `TOO_MANY_REQUESTS`, or `SERVICE_UNAVAILABLE`.",
            "example": "NOT_FOUND"
          },
          "detail": {
            "type": "string",
            "description": "What went wrong with this request, in English.",
            "example": "Meter not found"
          },
          "instance": {
            "type": "string",
            "description": "The path of the request that failed.",
            "example": "/v1/meters/zzzz"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code.",
            "example": 404
          },
          "title": {
            "type": "string",
            "description": "A short summary of the kind of error.",
            "example": "Resource not found"
          },
          "type": {
            "type": "string",
            "format": "uri",
            "description": "A URI identifying the kind of error. It is an identifier and does not resolve to a page.",
            "example": "https://qafiyah.com/errors/not-found"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Optional. A free key from https://qafiyah.com/developers raises the rate limit. A key the API does not recognize counts as no key."
      }
    }
  },
  "security": [
    {},
    {
      "apiKey": []
    }
  ],
...
}